requirements.txt
discord.py
python-dotenv
python-twitter